Create README.md
Repository of classifiers used in my research about classification of Candida spp. with Time Series Classifiers.

Abstract
Fungal infections are a constant concern in Intensive Care Units (ICUs) around the world. Candida spp. is one of the main reasons for these occurrences. Studies report that the mortality rate from infections caused by this category of microorganism is in the order of 40 to 60%. Currently, the gold standard technique for the laboratory diagnosis of candidemia is blood culture, but there are also other methods based on the analysis of Volatile Organic Compounds (VOCs), all of which require 2 to 7 days to obtain a result. Currently, a tool that is proving very promising in this context is the Electronic Nose (E-nose), which uses a set of gas sensors combined with Artificial Intelligence techniques to identify and classify VOCs. In this sense, this study aims to use an E-nose combined with classification techniques based on Time Series for the analysis and identification of Candida spp. in hospitalized patients, in order to reduce the decision-making time of health professionals regarding the most appropriate treatment. The main experiments show promising results, with high assertiveness rates for most models, with InceptionTime being the one with the best performance.